PMM-12522 Dropped new chunks collector and moved data to shards collector.
