+++ exporter/shards_collector.go
@@ -56,8 +56,29 @@ func (d *shardsCollector) collect(ch chan<- prometheus.Metric) {
 	client := d.base.client
 	logger := d.base.logger
 	prefix := "shards collection chunks"
+	ctx := d.ctx
 
-	// TODO: PMM-12522 find other places where we can use list databases and collections
+	metrics := make([]prometheus.Metric, 0)
+	logger.Debugf("chunksTotal")
+	metric, err := chunksTotal(ctx, client)
+	if err != nil {
+		logger.Debugf("cannot create metric for chunks total: %s", err)
+	} else {
+		metrics = append(metrics, metric)
+	}
+
+	logger.Debugf("chunksTotalPerShard")
+	ms, err := chunksTotalPerShard(ctx, client)
+	if err != nil {
+		logger.Debugf("cannot create metric for chunks total per shard: %s", err)
+	} else {
+		metrics = append(metrics, ms...)
+	}
+
+	for _, metric := range metrics {
+		ch <- metric
+	}
+
 	databaseNames, err := client.ListDatabaseNames(d.ctx, bson.D{})
 	if err != nil {
 		logger.Errorf("cannot get database names: %s", err)
@@ -185,6 +206,57 @@ func (d *shardsCollector) getChunksForCollection(row primitive.M) []bson.M {
 	}
 
 	return chunks
+}
+
+func chunksTotal(ctx context.Context, client *mongo.Client) (prometheus.Metric, error) {
+	n, err := client.Database("config").Collection("chunks").CountDocuments(ctx, bson.M{})
+	if err != nil {
+		return nil, err
+	}
+
+	name := "mongodb_mongos_sharding_chunks_total"
+	help := "Total number of chunks"
+
+	d := prometheus.NewDesc(name, help, nil, nil)
+	return prometheus.NewConstMetric(d, prometheus.GaugeValue, float64(n))
+}
+
+func chunksTotalPerShard(ctx context.Context, client *mongo.Client) ([]prometheus.Metric, error) {
+	aggregation := bson.D{
+		{Key: "$group", Value: bson.M{"_id": "$shard", "count": bson.M{"$sum": 1}}},
+	}
+
+	cursor, err := client.Database("config").Collection("chunks").Aggregate(ctx, mongo.Pipeline{aggregation})
+	if err != nil {
+		return nil, err
+	}
+
+	var shards []bson.M
+	if err = cursor.All(ctx, &shards); err != nil {
+		return nil, err
+	}
+
+	metrics := make([]prometheus.Metric, 0, len(shards))
+
+	for _, shard := range shards {
+		help := "Total number of chunks per shard"
+		labels := map[string]string{"shard": shard["_id"].(string)}
+
+		d := prometheus.NewDesc("mongodb_mongos_sharding_shard_chunks_total", help, nil, labels)
+		val, ok := shard["count"].(int32)
+		if !ok {
+			continue
+		}
+
+		metric, err := prometheus.NewConstMetric(d, prometheus.GaugeValue, float64(val))
+		if err != nil {
+			continue
+		}
+
+		metrics = append(metrics, metric)
+	}
+
+	return metrics, nil
 }
 
 var _ prometheus.Collector = (*shardsCollector)(nil)
